Update patch set 2
Patch Set 2: Code-Review-1

I don't quite get this patch.

This patch purports to add a dependency for ARM64, but dlopen already searches the LD_PATH, which would include system/lib64, system/vendor/lib64, etc.

Also, LOCAL_SHARED_LIBRARIES is meant for libraries which are to be linked against, not dlopen'ed.

The commit message is misleading, and the use of LOCAL_SHARED_LIBRARIES is not the correct way to handle this.

Patch-set: 2
Label: Code-Review=-1